Exclude any pull requests on master forks from release nightly
It seems pull request CI runs where included in the nighly build if the pull req originated from a master branch.
We do not want this behaviour.
The only way to filter on all types of events that kicked of the CI that we want seem to be to download and then filter locally